2539: revamp based on design review, to match with function injection. extracted shared component and removed duplicate code.
