+++ src/flows/components/Sidebar.scss
@@ -66,9 +66,3 @@
 .flow-sidebar--secrets-list {
   margin-right: -16px;
 }
-
-.flow-sidebar--secrets-list--toolbar {
-  margin-block-start: $cf-space-xs;
-  margin-block-end: $cf-space-2xs;
-  margin-right: 16px;
-}
