+++ src/flows/pipes/RawFluxEditor/functions.tsx
@@ -3,8 +3,9 @@ import React, {FC, useState, useMemo, useCallback} from 'react'
 import {EmptyState, ComponentSize} from '@influxdata/clockface'
 import {FLUX_FUNCTIONS} from 'src/shared/constants/fluxFunctions'
 import {FluxToolbarFunction} from 'src/types/shared'
-import Fn from './function'
+import Fn from 'src/flows/pipes/RawFluxEditor/FluxInjectionOption'
 import SearchWidget from 'src/shared/components/search_widget/SearchWidget'
+import FunctionTooltipContent from 'src/flows/pipes/RawFluxEditor/FunctionToolTipContent'
 
 interface Props {
   onSelect: (fn: FluxToolbarFunction) => void
@@ -56,10 +57,11 @@ const Functions: FC<Props> = ({onSelect}) => {
           <dt className="flux-toolbar--heading">{category}</dt>
           {fns.map(fn => (
             <Fn
-              onClickFunction={onSelect}
+              onClick={onSelect}
               key={`${fn.name}_${fn.desc}`}
-              func={fn}
+              option={fn}
               testID={fn.name}
+              ToolTipContent={FunctionTooltipContent}
             />
           ))}
         </dl>
